Removed MiniAppView and moved the code to RealMiniAppView
