+++ MiniApp/Classes/Display/Displayer.swift
@@ -1,6 +1,6 @@
 internal class Displayer {
 
     func getMiniAppView(miniAppId: String) -> MiniAppDisplayProtocol? {
-        return RealMiniAppView.shared.getMiniAppView(miniAppId: miniAppId)
+        return RealMiniAppView(miniAppId: miniAppId)
     }
 }
